refactor: extract TravelSidebar to own component
